wrong type cast.

+++ cbits/raaz/core/endian.c
@@ -3,8 +3,8 @@
 /* Include the right header file based on the platform */
 #ifdef __GNUC__
 #include <byteswap.h>
-uint32_t raazSwap32(uint32_t a){ return  bswap_32(a);}
-uint32_t raazSwap64(uint32_t a){ return bswap_64(a);}
+uint32_t raazSwap32(uint32_t a){ return bswap_32(a);}
+uint64_t raazSwap64(uint32_t a){ return bswap_64(a);}
 
 void raazSwap32Array(uint32_t *ptr, int n)
 {
